+++ e-app/src/lib.rs
@@ -9,6 +9,7 @@ mod dialogs;
 mod recent;
 mod git_view;
 mod builtin_completion;
+mod framework_completion;
 mod about;
 mod update_view;
 mod updater;